Type check correction + moving to workflow utils
